requirements.txt
# Core PDF processing
PyMuPDF>=1.23.0

# OCR support
pytesseract>=0.3.10
opencv-python>=4.8.0
Pillow>=10.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Networking and APIs
requests>=2.31.0

# Progress bars
tqdm>=4.65.0

# BibTeX processing
bibtexparser>=1.4.0

# Text processing
python-docx>=0.8.11

# Optional dependencies for enhanced functionality
# Uncomment if needed:
# scikit-image>=0.21.0  # For advanced image processing
# matplotlib>=3.7.0     # For visualization
# seaborn>=0.12.0       # For enhanced plotting
